README: Add badges to CI, crates.io, docs.rs and show the MSRV
I was trying to quickly get to the documentation of this crate and had
the GitHub page open... but there was no link on the front-page: let's
fix that.

+++ android-activity/src/input.rs
@@ -63,7 +63,7 @@ pub enum Class {
 
 impl From<u32> for Class {
     fn from(source: u32) -> Self {
-        let class = SourceFlags::from_bits_truncate(source as u32);
+        let class = SourceFlags::from_bits_truncate(source);
         match class {
             SourceFlags::BUTTON => Class::Button,
             SourceFlags::POINTER => Class::Pointer,
